+++ src/Knit/Effect/RandomFu.hs
@@ -13,6 +13,7 @@
 {-# LANGUAGE AllowAmbiguousTypes           #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving    #-}
 {-# LANGUAGE InstanceSigs                  #-}
+{-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -fwarn-incomplete-patterns #-}
 {-|
 Module      : Knit.Effect.RandomFu
@@ -33,16 +34,22 @@ This can be run in a few ways:
 module Knit.Effect.RandomFu
   (
     -- * Effect
-    Random
+    RandomFu
 
     -- * Actions
   , sampleRVar
   , sampleDist
 
   -- * Interpretations 
-  , runRandomIOSimple
-  , runRandomIOPureMT
-  , runRandomFromSource
+  , runRandomFuIOSimple
+  , runRandomFuIOPureMT
+  , runRandomFuFromSource
+
+  -- * Interop  
+  , absorbMonadRandom
+
+  -- * Deprecated, will be removed in next release
+  , Random
   )
 where
 
@@ -60,69 +67,73 @@ import           Control.Monad.IO.Class         ( MonadIO(..) )
 --import           Data.Kind                      ( Constraint )
 
 -- | Random Effect
-data Random m r where
-  SampleRVar ::  R.RVar t -> Random m t
-  GetRandomPrim :: R.Prim t -> Random m t
+data RandomFu m r where
+  SampleRVar ::  R.RVar t -> RandomFu m t
+  GetRandomPrim :: R.Prim t -> RandomFu m t
+
+type Random = RandomFu
+{-# DEPRECATED Random "Use RandomFu instead" #-}
 
 -- | Convert a random-fu RVar to the Random Effect
-sampleRVar :: (P.Member Random effs) => R.RVar t -> P.Sem effs t
+sampleRVar :: (P.Member RandomFu effs) => R.RVar t -> P.Sem effs t
 sampleRVar = send . SampleRVar
 
 -- | Convert a random-fu Distribution to the Random Effect
-sampleDist :: (P.Member Random effs, R.Distribution d t) => d t -> P.Sem effs t
+sampleDist
+  :: (P.Member RandomFu effs, R.Distribution d t) => d t -> P.Sem effs t
 sampleDist = sampleRVar . R.rvar
 
-getRandomPrim :: P.Member Random effs => R.Prim t -> P.Sem effs t
+getRandomPrim :: P.Member RandomFu effs => R.Prim t -> P.Sem effs t
 getRandomPrim = send . GetRandomPrim
 
 -- | Run in IO using default random-fu IO source
-runRandomIOSimple
+runRandomFuIOSimple
   :: forall effs a
    . MonadIO (P.Sem effs)
-  => P.Sem (Random ': effs) a
+  => P.Sem (RandomFu ': effs) a
   -> P.Sem effs a
-runRandomIOSimple = P.interpret f
+runRandomFuIOSimple = P.interpret f
  where
-  f :: forall m x . (Random m x -> P.Sem effs x)
+  f :: forall m x . (RandomFu m x -> P.Sem effs x)
   f r = case r of
     SampleRVar    rv -> liftIO $ R.sample rv
     GetRandomPrim pt -> liftIO $ R.getRandomPrim pt
 
 -- | Run using the given source
-runRandomFromSource
+runRandomFuFromSource
   :: forall s effs a
    . R.RandomSource (P.Sem effs) s
   => s
-  -> P.Sem (Random ': effs) a
+  -> P.Sem (RandomFu ': effs) a
   -> P.Sem effs a
-runRandomFromSource source = P.interpret f
+runRandomFuFromSource source = P.interpret f
  where
-  f :: forall m x . (Random m x -> P.Sem effs x)
+  f :: forall m x . (RandomFu m x -> P.Sem effs x)
   f r = case r of
     SampleRVar    rv -> R.runRVar (R.sample rv) source
     GetRandomPrim pt -> R.runRVar (R.getRandomPrim pt) source
 
 -- | Run in 'IO', using the given 'PureMT' source stored in an 'IORef'
-runRandomIOPureMT
+runRandomFuIOPureMT
   :: MonadIO (P.Sem effs)
   => R.PureMT
-  -> P.Sem (Random ': effs) a
+  -> P.Sem (RandomFu ': effs) a
   -> P.Sem effs a
-runRandomIOPureMT source re =
-  liftIO (newIORef source) >>= flip runRandomFromSource re
+runRandomFuIOPureMT source re =
+  liftIO (newIORef source) >>= flip runRandomFuFromSource re
 
--- | supply instance of MonadRandom for functions which require it
-$(R.monadRandom [d|
-        instance P.Member Random effs => R.MonadRandom (P.Sem effs) where
-            getRandomPrim = getRandomPrim
-    |])
-
-{- TODO
-newtype RandomSem r a = RandomSem { unRandomSem :: P.Sem r a } deriving (Functor, Applicative, Monad)
+newtype RandomFuSem r a = RandomFuSem { unRandomFuSem :: P.Sem r a } deriving (Functor, Applicative, Monad)
 
 $(R.monadRandom [d|
-        instance P.Member Random r => R.MonadRandom (RandomSem r) where
-            getRandomPrim :: R.Prim t -> RandomSem r t
-            getRandomPrim = RandomSem . getRandomPrim
+        instance P.Member RandomFu r => R.MonadRandom (RandomFuSem r) where
+            getRandomPrim = RandomFuSem . getRandomPrim
     |])
+
+{- | Given a function that uses the random-fu package and produces a result thusly
+constrained by 'MonadRandom', absorb it into a Polysemy monad whose
+effect list contains the RandomFu effect.
 -}
+absorbMonadRandom
+  :: P.Member RandomFu r => (forall m . R.MonadRandom m => m a) -> P.Sem r a
+absorbMonadRandom = unRandomFuSem
+